+++ packages/adapter-components/src/client/index.ts
@@ -19,5 +19,7 @@ export { DEFAULT_RETRY_OPTS, RATE_LIMIT_UNLIMITED_MAX_CONCURRENT_REQUESTS } from
 export { logDecorator, requiresLogin } from './decorators'
 export { AdapterHTTPClient, ClientBaseParams, ClientDataParams, ClientOpts, HTTPReadClientInterface, HTTPWriteClientInterface, HTTPError, HttpMethodToClientParams } from './http_client'
 export { axiosConnection, createClientConnection, createRetryOptions, validateCredentials, APIConnection, ConnectionCreator, Response, ResponseValue, UnauthorizedError, Connection, RetryOptions, AuthParams, AuthenticatedAPIConnection } from './http_connection'
-export { createPaginator, getWithCursorPagination, getWithItemOffsetPagination as getWithItemIndexPagination, getWithPageOffsetPagination, getWithPageOffsetAndLastPagination, getAllPagesWithOffsetAndTotal, traverseRequests, ClientGetWithPaginationParams, GetAllItemsFunc, PageEntriesExtractor, PaginationFunc, PaginationFuncCreator, Paginator, PathCheckerFunc, defaultPathChecker } from './pagination'
+export { createPaginator, getWithCursorPagination, getWithItemOffsetPagination as getWithItemIndexPagination, getWithPageOffsetPagination, getWithPageOffsetAndLastPagination, traverseRequests, PathCheckerFunc, defaultPathChecker } from './pagination/pagination'
+export { ClientGetWithPaginationParams, GetAllItemsFunc, PageEntriesExtractor, PaginationFunc, PaginationFuncCreator, Paginator } from './pagination/common'
+export { getAllPagesWithOffsetAndTotal } from './pagination/pagination_async'
 export { createRateLimitersFromConfig, throttle, BottleneckBuckets } from './rate_limit'
